abstract profile logic from widget_tree into profile page
